Fix issue with computed Median
Looking at the code for Median and the Quantile function it points to, I think the issue is that n is determined at the start of the function (line 497) prior to detection/removal of NA values. I've added a line replacing the n after removing the NA values. 

I think this would close #67